feat: split zksync hardhat config
