+++ package.json
@@ -9,15 +9,12 @@
     "cross chain"
   ],
   "devDependencies": {
-    "@matterlabs/hardhat-zksync-deploy": "^0.6.3",
-    "@matterlabs/hardhat-zksync-solc": "^0.3.16",
-    "@matterlabs/hardhat-zksync-verify": "^0.1.5",
-    "@matterlabs/hardhat-zksync-upgradable": "^0.1.3",
-    "@nomicfoundation/hardhat-toolbox": "^2.0.2",
+    "@matterlabs/hardhat-zksync-deploy": "^1.1.0",
+    "@nomicfoundation/hardhat-toolbox": "^4.0.0",
     "@nomiclabs/hardhat-solpp": "^2.0.1",
-    "hardhat": "^2.13.1",
-    "zksync-web3": "^0.14.3",
-    "@openzeppelin/hardhat-upgrades": "^1.27.0"
+    "@openzeppelin/hardhat-upgrades": "^3.0.0",
+    "hardhat": "^2.19.2",
+    "zksync-ethers": "^6.0.0"
   },
   "scripts": {
     "compile-master": "MASTER_UNITTEST=true npx hardhat compile",
