+++ script/deploy_l1_gateway.js
@@ -47,14 +47,15 @@ task("deployL1Gateway", "Deploy L1 Gateway")
       if (!(logName.DEPLOY_GATEWAY_TARGET in deployLog) || force) {
         console.log("start deploy contract");
         const contract = await ethers.getContractFactory(contractName);
-        instance = await upgrades.deployProxy(contract, initializeParams, {
+        const tx = await upgrades.deployProxy(contract, initializeParams, {
           kind: "uups",
         });
+        instance.address = await tx.getAddress()
         deployLog[logName.DEPLOY_GATEWAY] = instance.address;
         console.log("instance address and wait deployed:", instance.address);
         fs.writeFileSync(deployLogPath, JSON.stringify(deployLog));
 
-        const receipt = await instance.deployTransaction.wait()
+        const receipt = await tx.waitForDeployment()
         deployLog[logName.DEPLOY_LOG_DEPLOY_TX_HASH] = receipt.transactionHash
         deployLog[logName.DEPLOY_LOG_DEPLOY_BLOCK_NUMBER] = receipt.blockNumber
         console.log("deployed success:", instance.address);
